updated version, fixed issue with rendering dependency graph component
