+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@backstage-plugin-progressive-delivery/backstage-plugin-progressive-delivery-frontend",
-  "version": "0.1.3",
+  "version": "0.2.0",
   "main": "dist/index.esm.js",
   "types": "dist/index.d.ts",
   "license": "Apache-2.0",
